+++ src/components/settings/menu/settingsLogo.tsx
@@ -1,9 +1,9 @@
-import '../../../features/logoTweaks.css';
+import '../../../features/logoVariations.css';
 
 import {css} from '@emotion/css';
 import React, {FC} from 'react';
 
-import {BTDLogoVariations, BTDStandaloneLogo} from '../../../features/logoTweaks';
+import {BTDLogoVariations, BTDStandaloneLogo} from '../../../features/logoVariations';
 import {getTransString} from '../../trans';
 import {NewFeatureBadge} from '../components/newFeatureBadge';
 import {BTDRadioSelectSettingsRow} from '../components/radioSelectSettingsRow';
