--- src/features/logoTweaksVariables.css
@@ -1,74 +1,0 @@
-:root {
-  --btd-logo-background: url('../assets/logo-tweaks/btd.svg');
-  --btd-logo-border-color: transparent;
-}
-
-[btd-logo='asexual'] {
-  --btd-logo-background: url('../assets/logo-tweaks/asexual.svg');
-}
-[btd-logo='agender'] {
-  --btd-logo-background: url('../assets/logo-tweaks/agender.svg');
-}
-[btd-logo='androgyne'] {
-  --btd-logo-background: url('../assets/logo-tweaks/androgyne.svg');
-}
-[btd-logo='aromantic'] {
-  --btd-logo-background: url('../assets/logo-tweaks/aromantic.svg');
-}
-[btd-logo='bigender'] {
-  --btd-logo-background: url('../assets/logo-tweaks/bigender.svg');
-}
-[btd-logo='bisexual'] {
-  --btd-logo-background: url('../assets/logo-tweaks/bisexual.svg');
-}
-[btd-logo='demigirl'] {
-  --btd-logo-background: url('../assets/logo-tweaks/demigirl.svg');
-}
-[btd-logo='demiguy'] {
-  --btd-logo-background: url('../assets/logo-tweaks/demiguy.svg');
-}
-[btd-logo='deminonbinary'] {
-  --btd-logo-background: url('../assets/logo-tweaks/deminonbinary.svg');
-}
-[btd-logo='demisexual'] {
-  --btd-logo-background: url('../assets/logo-tweaks/demisexual.svg');
-}
-[btd-logo='enbian'] {
-  --btd-logo-background: url('../assets/logo-tweaks/enbian.svg');
-}
-[btd-logo='genderfluid'] {
-  --btd-logo-background: url('../assets/logo-tweaks/genderfluid.svg');
-}
-[btd-logo='genderqueer'] {
-  --btd-logo-background: url('../assets/logo-tweaks/genderqueer.svg');
-}
-[btd-logo='intersex'] {
-  --btd-logo-background: url('../assets/logo-tweaks/intersex.svg');
-}
-[btd-logo='lesbian'] {
-  --btd-logo-background: url('../assets/logo-tweaks/lesbian.svg');
-}
-[btd-logo='neutrois'] {
-  --btd-logo-background: url('../assets/logo-tweaks/neutrois.svg');
-}
-[btd-logo='non binary'] {
-  --btd-logo-background: url('../assets/logo-tweaks/non binary.svg');
-}
-[btd-logo='omnisexual'] {
-  --btd-logo-background: url('../assets/logo-tweaks/omnisexual.svg');
-}
-[btd-logo='pansexual'] {
-  --btd-logo-background: url('../assets/logo-tweaks/pansexual.svg');
-}
-[btd-logo='polyamory'] {
-  --btd-logo-background: url('../assets/logo-tweaks/polyamory.svg');
-}
-[btd-logo='polysexual'] {
-  --btd-logo-background: url('../assets/logo-tweaks/polysexual.svg');
-}
-[btd-logo='rainbow'] {
-  --btd-logo-background: url('../assets/logo-tweaks/rainbow.svg');
-}
-[btd-logo='trans'] {
-  --btd-logo-background: url('../assets/logo-tweaks/trans.svg');
-}
