+++ src/features/logoVariations.tsx
@@ -1,4 +1,4 @@
-import './logoTweaks.css';
+import './logoVariations.css';
 
 import React from 'react';
 import {renderToStaticMarkup} from 'react-dom/server';
