+++ src/inject.ts
@@ -19,7 +19,7 @@ import {maybeFreezeGifsInProfilePicture} from './features/freezeGifsProfilePictu
 import {setupGifModals} from './features/gifModals';
 import {maybeHideColumnIcons} from './features/hideColumnIcons';
 import {keepTweetedHashtagsInComposer} from './features/keepTweetedHashtags';
-import {changeLogo} from './features/logoTweaks';
+import {changeLogo} from './features/logoVariations';
 import {makeSearchColumnsFirst} from './features/makeSearchColumnsFirst';
 import {useModernOverlays} from './features/modernOverlays';
 import {pauseColumnsOnHover} from './features/pauseColumnsOnHover';
